Add Malay and Malay (Jawi) translations
Sorry for not noticing the new translation keys... Been quite busy with a lot of stuff so I have not checked GitHub for quite a while now.

This PR now includes the new translation keys, and also a minor changes to one of the existing translation I submitted back then.